Clean up after code review.
