+++ bindings/wasm/bin/manifold-cad
@@ -19,35 +19,61 @@ import { resolve, extname } from 'node:path';
 import { Command } from 'commander';
 import glMatrix from 'gl-matrix';
 
-import { Evaluator } from '../lib/evaluate.js';
+import * as animation from '../lib/animation.js';
+import * as debug from '../lib/debug.js';
+import {Evaluator} from '../lib/evaluate.js';
 import * as exporter from '../lib/export.js';
+import {Export3MF} from '../lib/export-3mf.js';
+import {ExportGLTF} from '../lib/export-gltf.js';
+import * as material from '../lib/material.js';
 
 // Setup the evaluator and context.
 const evaluator = new Evaluator();
-evaluator.addContext({glMatrix})
+evaluator.addContext({glMatrix});
 
 // These are exporter methods that generate Manifold
 // or CrossSection objects.  Tell the evaluator to intercept
 // the calls, and add any created objects to the clean up list.
-for (const name of ['show', 'only', 'setMaterial']) {
-  evaluator.addContextMethodWithCleanup(name, exporter[name]);
-}
+evaluator.addContextMethodWithCleanup('show', debug.show)
+evaluator.addContextMethodWithCleanup('only', debug.only)
+evaluator.addContextMethodWithCleanup('setMaterial', material.setMaterial)
 
 // Add additional exporter context.  These need no garbage collection.
 evaluator.addContext({
   GLTFNode: exporter.GLTFNode,
-  setMorphStart: exporter.setMorphStart,
-  setMorphEnd: exporter.setMorphEnd,
+  setMorphStart: animation.setMorphStart,
+  setMorphEnd: animation.setMorphEnd,
 });
 
-const evaluateCADToModel = async (code) => {
+// Instantiate exporters for later use.
+const exporters = [
+  new Export3MF(),
+  new ExportGLTF()
+];
+
+const evaluateCADToModel = (code) => {
+  // Global defaults can be populated by the script.  It's set per
+  // evaluation, while the rest of evaluator context doesn't change from
+  // run to run.
+  // This can be used to set parameters elsewhere in ManifoldCAD.  For
+  // example, the GLTF exporter will look for animation type and
+  // framerate.
   const globalDefaults = {};
 
+  // Evaluate the code
   evaluator.context.globalDefaults = globalDefaults;
   const manifold = evaluator.evaluate(code);
 
-  return await exporter.exportModels(globalDefaults, manifold);
-}
+  // If we don't actually have a model, complain.
+  if (!manifold && !exporter.hasGLTFNodes()) {
+    console.error('No output because "result" is undefined and no "GLTFNode"s were created.');
+    process.exit(2);
+  }
+
+  // Create a gltf-transform document.
+  return exporter.hasGLTFNodes() ?
+      exporter.GLTFNodesToGLTFDoc(exporter.getGLTFNodes(), globalDefaults) :
+      exporter.manifoldToGLTFDoc(manifold, globalDefaults);}
 
 const program = new Command();
 
@@ -60,17 +86,21 @@ program
     const inPath = resolve(infile);
     const outPath = resolve(outfile);
 
-    const is3mf = extname(outPath) === '.3mf';
-    const isGlb = extname(outPath) === '.glb';
-    if (!is3mf && !isGlb) {
-      throw new Error('Output file must be either .glb or .3mf');
+    // Choose an exporter.
+    const exporter = exporters.find(ex => ex.extensions.includes(extname(outPath)));
+    if (!exporter) {
+      const extensions = exporters.map(ex => ex.extensions.join(', ')).join(', ');
+      console.error(`Output file must be one of ${extensions}`);
+      process.exit(1);
     }
 
+    // Build the model.
     const code = await readFile(inPath);
-    const { glbBlob, threeMFBlob} = await evaluateCADToModel(code);
+    const doc = evaluateCADToModel(code);
 
-    const binary = is3mf ? threeMFBlob : glbBlob;
-    const outStream = writeFile(outfile, binary.stream());
+    // Convert it and write it to disk.
+    const binary = await exporter.asBlob(doc);
+    await writeFile(outfile, binary.stream());
   })
 
 program.parseAsync(process.argv);
