--- bindings/wasm/lib/.gitignore
@@ -1,3 +1,0 @@
-*.js
-*.js.map
-*.d.ts
