+++ bindings/wasm/lib/animation.ts
@@ -1,13 +1,26 @@
+// Copyright 2022-2025 The Manifold Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import {Accessor, Animation, AnimationSampler, Document, Mesh as GLTFMesh, Node} from '@gltf-transform/core';
 import {quat} from 'gl-matrix';
 
 import {Manifold, Mesh, Vec3} from '../examples/built/manifold';
 import {Quat} from '../examples/public/editor';
 
-import {globalDefaults, GLTFNode} from './export'
+import {globalDefaults, GLTFNode} from './export.ts'
 
 const FPS = 30;
-
 
 interface Morph {
   start?: (v: Vec3) => void;
