+++ bindings/wasm/lib/material.ts
@@ -1,11 +1,24 @@
+// Copyright 2022-2025 The Manifold Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import {Document, Material} from '@gltf-transform/core';
 import {KHRMaterialsUnlit} from '@gltf-transform/extensions';
 
-import {Manifold} from '../examples/built/manifold';
-import {GLTFMaterial} from '../examples/public/editor';
+import {Manifold} from '../examples/built/manifold.ts';
+import {GLTFMaterial} from '../examples/public/editor.ts';
 
-import {globalDefaults, GLTFNode} from './export'
+import {globalDefaults, GLTFNode} from './export.ts'
 
 const id2material = new Map<number, GLTFMaterial>();
 const materialCache = new Map<GLTFMaterial, Material>();
