+++ bindings/wasm/package.json
@@ -9,8 +9,7 @@
   "scripts": {
     "docs": "npx typedoc",
     "build": "tsc",
-    "prepare": "npm run build",
-    "clean": "rm -f lib/*.js lib/*.js.map lib/*.d.ts lib/gltf-transform/*.js lib/gltf-transform/*.js.map lib/gltf-transform/export-3mf.d.ts lib/gltf-transform/manifold-gltf.d.ts"
+    "prepare": "npm run build"
   },
   "files": [
     "manifold.js",
@@ -19,17 +18,25 @@
     "manifold-encapsulated-types.d.ts",
     "manifold-global-types.d.ts",
     "bin/manifold-cad",
+    "lib/animation.d.ts",
+    "lib/animation.js",
+    "lib/debug.d.ts",
+    "lib/debug.js",
     "lib/evaluate.d.ts",
     "lib/evaluate.js",
     "lib/export.d.ts",
     "lib/export.js",
     "lib/gltf-io.d.ts",
     "lib/gltf-io.js",
-    "lib/gltf-transform/manifold-gltf.d.ts",
-    "lib/gltf-transform/manifold-gltf.js",
-    "lib/gltf-transform/3mf-export.d.ts",
-    "lib/gltf-transform/export3mf.d.ts",
-    "lib/gltf-transform/export3mf.js"
+    "lib/3mf-export.d.ts",
+    "lib/export-3mf.d.ts",
+    "lib/export-3mf.js",
+    "lib/export-gltf.d.ts",
+    "lib/export-gltf.js",
+    "lib/manifold-gltf.d.ts",
+    "lib/manifold-gltf.js",
+    "lib/material.d.ts",
+    "lib/material.js"
   ],
   "typings": "manifold.d.ts",
   "types": "manifold.d.ts",
