+++ bindings/wasm/tsconfig.json
@@ -20,8 +20,7 @@
     "rewriteRelativeImportExtensions": true,
   },
   "include": [
-    "./lib/*.ts",
-    "lib/gltf-transform/*.ts"
+    "./lib/*.ts"
   ],
   "exclude": []
 }
