+++ bindings/wasm/typedoc.json
@@ -4,8 +4,7 @@
     "manifold-encapsulated-types.d.ts",
     "manifold-global-types.d.ts",
     "lib/evaluate.ts",
-    "lib/gltf-transform/manifold-gltf.ts",
-    "lib/gltf-transform/export-3mf.ts"
+    "lib/manifold-gltf.ts"
   ],
   "sortEntryPoints": false,
   "projectDocuments": ["documents/*.md"],
